+++ CHANGELOG.md
@@ -1,5 +1,16 @@
-# LightConeCalc change log
+# Cosmic expansion change log
 
-## `1.0.0` _2022-07-12_
+## ## 1.1.0 _2022-07-24_
+
+### Added
+
+- More unit conversions for external use.
+- Calculate omegaLambda0 from omegaM0 unless it is provided.
+
+### Fixed
+
+- Correct duplication of z = 0 if at the start or end of the range.
+
+## 1.0.0 _2022-07-12_
 
 Import first release.